+++ index.html
@@ -235,5 +235,5 @@
     </template>
 
     <script src="/js/functions.js"></script>
-    <script src="/js/main.js"></script>
+    <script src="/js/main.js" type="module" ></script>
 </html>
